Move `compare.ts` out of the dead `resource` folder
