+++ src/devtools/client/debugger/src/actions/sources/index.ts
@@ -2,7 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
-//
 export * from "./blackbox";
 export * from "./newSources";
 export * from "./select";
