+++ src/devtools/client/debugger/src/utils/compare.ts
